feat: sonarqube 적용

+++ src/main/java/com/monew/monew_server/domain/interest/service/InterestService.java
@@ -121,9 +121,9 @@ public class InterestService {
     @Transactional
     public SubscriptionDto subscribe(UUID interestId, UUID userId) {
 
-        userRepository.findById(userId).orElseThrow(() ->
-            new InterestException(ErrorCode.USER_NOT_FOUND)
-        );
+        if (!userRepository.existsById(userId)) {
+            throw new InterestException(ErrorCode.USER_NOT_FOUND);
+        }
 
         Interest interest = interestRepository.getOrThrow(interestId);
 
@@ -153,9 +153,9 @@ public class InterestService {
     @Transactional
     public void unsubscribe(UUID interestId, UUID userId) {
 
-        userRepository.findById(userId).orElseThrow(() ->
-            new InterestException(ErrorCode.USER_NOT_FOUND)
-        );
+        if (!userRepository.existsById(userId)) {
+            throw new InterestException(ErrorCode.USER_NOT_FOUND);
+        }
 
         interestRepository.getOrThrow(interestId);
         subscriptionRepository.deleteByUserIdAndInterestId(userId, interestId);
